requirements.txt
Flask==2.2.5
gunicorn
# Azure Managed Identity
azure-identity
# Azure App Configuration
azure-appconfiguration
# App Insights
opencensus
opencensus-ext-flask
opencensus-ext-azure

